+++ src/devices/DeviceBase.ts
@@ -56,8 +56,8 @@ export abstract class DeviceBase {
   }
 
   async getDeviceLogSettings(device: devicesConfig): Promise<void> {
-    this.deviceLogging = this.platform.debugMode ? 'debugMode' : device.logging ?? this.platform.platformLogging ?? 'standard'
-    const logging = this.platform.debugMode ? 'Debug Mode' : device.logging ? 'Device Config' : this.platform.platformLogging ? 'Platform Config' : 'Default'
+    this.deviceLogging = device.logging ?? this.platform.platformLogging ?? 'standard'
+    const logging = device.logging ? 'Device Config' : this.platform.platformLogging ? 'Platform Config' : 'Default'
     await this.debugLog(`Using ${logging} Logging: ${this.deviceLogging}`)
   }